refactor: move index code to run
This makes the code a bit more predicatable, it only self-executes when importing index
